+++ components/ui/bot-status.tsx
